tests: fs: Move tests to the common directory
There are a lot of tests in the LittleFS tests directory that can be
used in different file systems as well. These tests are moved to common
directory.
