+++ tests/subsys/fs/littlefs/CMakeLists.txt
@@ -5,6 +5,14 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(littlefs)
 
 FILE(GLOB app_sources src/*.c)
+FILE(GLOB common_sources
+  ../common/test_fs_util.c
+  ../common/test_fs_open_flags.c
+  ../common/test_fs_mount_flags.c
+  ../common/test_fs_basic.c
+  ../common/test_fs_dirops.c
+)
+
 # LittleFS opens files for RW by default if no flags given and crashes,
 # instead of returning error when attempting to perform operation that
 # a file has not been opened for.
@@ -12,4 +20,7 @@ target_compile_definitions(app PRIVATE
 		BYPASS_FS_OPEN_FLAGS_LFS_ASSERT_CRASH
 		BYPASS_FS_OPEN_FLAGS_LFS_RW_IS_DEFAULT
 )
-target_sources(app PRIVATE ${app_sources} ../common/test_fs_open_flags.c)
+target_sources(app PRIVATE
+  ${app_sources}
+  ${common_sources}
+)
