+++ tests/subsys/fs/littlefs/src/testfs_tests.h
@@ -28,6 +28,6 @@ void test_lfs_perf(void);
 void test_fs_open_flags_lfs(void);
 
 /* Test fs_mount flags */
-void test_fs_mount_flags(void);
+void test_fs_mount_flags_lfs(void);
 
 #endif /* _ZEPHYR_TESTS_SUBSYS_FS_LITTLEFS_TESTFS_TESTS_H_ */
